Permit choice of dash to separate words in page and post names.
A word_sep keyword argument is added to several functions and is
ultimately used in utils.name_from_title(). I also ensure that only
one instance of the word sep char appears between words and that
there are no leading or trailing non-word chars in the name. I'd
like to propose the same for other page/post/draft creation options.

I just found Tinkerer yesterday and it seems to be exactly what I've
been looking for in a blogging tool!