Limit RSS feed contents and remove headerlinks
Add a new option, rss_max_items, to control the number of posts that
appear in RSS feeds.

Make the code for building the feed context more generic, based on the
work in tinkerer-contrib/rss/rss.py.

Strip anchor tags with the class headerlink from RSS item descriptions,
so when the links are used in a site they do not appear in the RSS
output where they don't make sense as permalinks.

Change-Id: Ia90d70389152a9fda1137606eb6eef9918cbc221